generate-#19/requirements.txt
python-pptx>=0.6.21
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
pdf2image>=3.1.0
PyMuPDF>=1.23.0
scikit-image>=0.20.0
comtypes>=1.2.0